Create Thomas Enechi
Thomas is a Co-Founder, ICO Investor and a Blockchain Advisor. He holds a M.Sc. in Information Technology, a First Degree in Electrical and Electronics Engineering as well as a Telecoms Mini-MBA from Informa Telecoms and Technology Academy, United Kingdom. He is a Blockchain enthusiast and a Cryptocurrency investor, with about 15 years of extensive experience in the Telecommunication and Information Technology industries both in Nigeria and abroad.

Thomas is the Chief Operating Officer of LIFTIX LLC; a Blockchain StartUp based in Lagos, Nigeria, owners of the VITR Identity Verification Network (VTK). A Blockchain Advisor to ENDO Blockchain Protocol based in Singapore, Member of the Board of Directors at International Decentralized Association of Cryptocurrency and Blockchain — IDACB and Country Manager, European Chamber of Distributed Ledger Technologies and Cryptocurrencies — EUCC.

Thomas’ vast experience, his advice and assistance will be an exceptional benefit for BEATRĪX and will be a great aid in merging more and more with the African community as well as the crypto-currency world. We are delighted to have you on board.

LinkedIn Profil: https://www.linkedin.com/in/chitom/
https://ibb.co/jOkpLz